+++ native/qt6/src/preview_overlay.h
@@ -14,7 +14,17 @@
 #include <QGraphicsView>
 #include <QGraphicsScene>
 #include <QGraphicsPixmapItem>
+#include <QGraphicsSvgItem>
 #include <QComboBox>
+#include <QCheckBox>
+#ifdef HAVE_QT_PDF
+#include <QPdfDocument>
+#include <QPdfView>
+#endif
+#include <QPlainTextEdit>
+#include <QTableView>
+#include <QStandardItemModel>
+
 #include "oiio_image_loader.h"
 
 class PreviewOverlay : public QWidget
@@ -30,6 +40,7 @@ public:
     void navigateNext();
     void navigatePrevious();
     void stopPlayback();
+    QString currentPath() const { return currentFilePath; }
 
 signals:
     void closed();
@@ -40,6 +51,7 @@ protected:
     void resizeEvent(QResizeEvent *event) override;
     void mousePressEvent(QMouseEvent *event) override;
     void wheelEvent(QWheelEvent *event) override;
+    bool eventFilter(QObject* watched, QEvent* event) override;
 
 private slots:
     void onPlayPauseClicked();
@@ -59,6 +71,14 @@ private:
     void setupUi();
     void showImage(const QString &filePath);
     void showVideo(const QString &filePath);
+#ifdef HAVE_QT_PDF
+    void showPdf(const QString &filePath);
+    void renderPdfPageToImage();
+#endif
+    void showDoc(const QString &filePath);
+    void showDocx(const QString &filePath);
+    void showXlsx(const QString &filePath);
+    void showText(const QString &filePath);
     void updatePlayPauseButton();
     QString formatTime(qint64 milliseconds);
     void zoomImage(double factor);
@@ -75,6 +95,7 @@ private:
     QGraphicsView *imageView;
     QGraphicsScene *imageScene;
     QGraphicsPixmapItem *imageItem;
+    QGraphicsSvgItem *svgItem;
     QVideoWidget *videoWidget;
     QWidget *controlsWidget;
     QPushButton *playPauseBtn;
@@ -85,10 +106,20 @@ private:
     QLabel *fileNameLabel;
     QComboBox *colorSpaceCombo;
     QLabel *colorSpaceLabel;
+    QCheckBox *alphaCheck;
+    QPlainTextEdit *textView;
+
+    QTableView *tableView;
+    QStandardItemModel *tableModel;
 
     // Media player
     QMediaPlayer *mediaPlayer;
     QAudioOutput *audioOutput;
+#ifdef HAVE_QT_PDF
+    QPdfDocument *pdfDoc;
+    QPdfView *pdfView;
+    int pdfCurrentPage = 0;
+#endif
 
     // State
     QString currentFilePath;
@@ -123,6 +154,10 @@ private:
     // Color space for HDR/EXR images
     OIIOImageLoader::ColorSpace currentColorSpace;
     bool isHDRImage;
+
+    // Alpha channel toggle state
+    bool alphaOnlyMode = false;
+    bool previewHasAlpha = false;
 };
 
 #endif // PREVIEW_OVERLAY_H