+++ native/qt6/src/settings_dialog.h
@@ -11,6 +11,9 @@
 #include <QCheckBox>
 #include <QGroupBox>
 #include <QTextEdit>
+#include <QTableWidget>
+#include <QKeySequenceEdit>
+#include <QLineEdit>
 
 class SettingsDialog : public QDialog {
     Q_OBJECT
@@ -31,26 +34,27 @@ private:
     void setupViewTab();
     void setupShortcutsTab();
     void setupAboutTab();
-    
+
     QTabWidget* tabWidget;
-    
+
     // General tab
     QComboBox* themeCombo;
-    
+
     // Cache tab
     QLabel* cacheSizeLabel;
     QPushButton* clearCacheBtn;
     QSpinBox* maxCacheSizeSpin;
-    
+
     // View tab
     QComboBox* viewModeCombo;
     QSpinBox* thumbnailSizeSpin;
     QCheckBox* showFileExtensionsCheck;
     QCheckBox* showSequenceOverlayCheck;
-    
-    // Shortcuts tab
-    QTextEdit* shortcutsText;
-    
+
+    // Shortcuts tab (editable)
+    QTableWidget* fmShortcutsTable;
+    QPushButton* fmResetAllBtn;
+
     // About tab
     QLabel* versionLabel;
     QLabel* qtVersionLabel;